workflows: migrated new logging structure
* Updated the new logging structure to work better
  overall and fit with the new code structure.
